requirements.txt
django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
channels==4.0.0
opencv-python==4.9.0.80
pyaudio==0.2.14
numpy==1.26.4
pillow==10.2.0
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
drf-yasg==1.21.7
python-jose==3.3.0
websockets==12.0
groq==0.4.1
transformers==4.37.2
torch==2.2.0 